EEH-2513 - Added json_validation.py for the GH actions
